Remove the MyProfile component confilict and Optimize that.
